Space X Falcon 9 First Stage Landing Prediction
Space X advertises Falcon 9 rocket launches on its website with a cost of 62 million dollars; other providers cost upward of 165 million dollars each, Much of the savings is because Space X can reuse the first stage. Therefore we to determine if the first stage will land, we can determine the cost of a launch.